Make minor fixes in GW150914.py
These are two minor fixes I found significantly improve the behavior of the example.
It still does not agree with standard results due to different priors
(and perhaps other things, but the priors are certainly different)
but the result is nonetheless much closer with these fixes.

The first is enabling the use of float64 by jax, which is necessary to prevent NaNs in the likelihood.
The second is to correct the keys in the transform dictionary, so that sin_dec and cos_iota will
be converted to dec and iota as they ought to be.